Fixed compiler warnings when warning for sign comparison is activated. Refactored singleton instances for controllers out of JMC class (saves memory, avoids weird stats when controller is used at different places in the UI). Added *WithMode methods to get view controller with or without navigation controller. Changed presentation of issues view controller to modal when notification is tapped and window.rootViewController is present.

+++ JIRAConnect/JMCClasses/Core/crashes/CrashReporter.m
@@ -177,7 +177,7 @@ static CrashReporter *crashReportSender = nil;
 
     NSFileManager *fm = [NSFileManager defaultManager];
 
-    for (int i = 0; i < [_crashFiles count]; i++)
+    for (NSUInteger i = 0; i < [_crashFiles count]; i++)
     {
         [fm removeItemAtPath:[_crashesDir stringByAppendingPathComponent:[_crashFiles objectAtIndex:i]] error:&error];
     }
@@ -190,7 +190,7 @@ static CrashReporter *crashReportSender = nil;
 
     NSMutableArray *crashReports = [NSMutableArray arrayWithCapacity:[_crashFiles count]];
 
-    for (int i = 0; i < [_crashFiles count]; i++)
+    for (NSUInteger i = 0; i < [_crashFiles count]; i++)
     {
         NSString *filename = [_crashesDir stringByAppendingPathComponent:[_crashFiles objectAtIndex:i]];
